Refactor # 66
Crear capa intermedia para el manejo de lógica y validaciones de negocio
