+++ paymentchainparent/businessdomain/customer/src/main/java/com/paymentchain/customer/controller/CustomerRestController.java
@@ -1,24 +1,12 @@
 package com.paymentchain.customer.controller;
 
 
-import java.time.Duration;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.TimeUnit;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import io.netty.channel.ChannelOption;
-import io.netty.channel.epoll.EpollChannelOption;
-import io.netty.handler.timeout.ReadTimeoutHandler;
-import io.netty.handler.timeout.WriteTimeoutHandler;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.client.reactive.ReactorClientHttpConnector;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,26 +16,17 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.reactive.function.client.WebClient;
-import reactor.netty.http.client.HttpClient;
 
 import com.paymentchain.customer.controller.Helpers.CustomerRESTControllerHelper;
 import com.paymentchain.customer.entities.Customer;
 import com.paymentchain.customer.entities.CustomerProduct;
 import com.paymentchain.customer.repository.CustomerRepository;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.env.Environment;
-
 
 @RestController
 @RequestMapping("/customer/v1")
 public class CustomerRestController {
     
-    private final String URL_PRODUCTS = "http://BUSINESSDOMAIN-PRODUCT/product";
-    private final String URL_TRANSACTIONS = "http://localhost:8083/transaction";
-    
-    @Autowired
-    private WebClient.Builder webClientBuilder;
     
     /*
     private final WebClient.Builder webClientBuilder;
@@ -57,16 +36,8 @@ public class CustomerRestController {
         this.webClientBuilder = aBuilder;
     }
     */
-    HttpClient client = HttpClient.create()
-            .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, 5000)
-            .option(ChannelOption.SO_KEEPALIVE, true)
-            .option(EpollChannelOption.TCP_KEEPIDLE, 300)
-            .option(EpollChannelOption.TCP_KEEPINTVL, 60)
-            .responseTimeout(Duration.ofSeconds(1))
-            .doOnConnected(connection -> {
-                connection.addHandlerLast(new ReadTimeoutHandler(5000, TimeUnit.MILLISECONDS));
-                connection.addHandlerLast(new WriteTimeoutHandler(5000, TimeUnit.MILLISECONDS));
-            });
+    
+    
     
     @Autowired
     CustomerRepository customerRepository;
@@ -147,56 +118,7 @@ public class CustomerRestController {
     
     @GetMapping("/full")
     public Customer getByCode(@RequestParam("code") String code) {
-         Customer customerByCode = customerRepository.findByCode(code);
-         List<CustomerProduct> customerProducts = customerByCode.getProducts();
-         customerProducts.forEach(p -> {
-             String productName = getProductName(p.getId());
-             p.setProductName(productName);
-         });
-         /*
-         String customerIBAN = customerByCode.getIBAN();
-         List<?> transactions = getTransactions(customerIBAN);
-         customerByCode.setTransactions(transactions);
-*/
-        return customerByCode;
-
+        return  CustomerRESTControllerHelper.getByCode(code);
     }
     
-    private String getProductName(long id) {
-        ReactorClientHttpConnector reactorClientHttpConnector = new ReactorClientHttpConnector(client);
-        WebClient build = webClientBuilder.clientConnector(reactorClientHttpConnector)
-                .baseUrl(URL_PRODUCTS)
-                .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                .defaultUriVariables(Collections.singletonMap("url", URL_PRODUCTS))
-        .build();
-        
-        JsonNode productResponse = build.method(HttpMethod.GET).uri("/" + id)
-                .retrieve().bodyToMono(JsonNode.class).block();
-        if (productResponse != null) {
-            return productResponse.get("name").asText();
-        }
-        
-        return "";
-    }
-    
-    
-    private List<?> getTransactions(String ibanAccount) {
-        ReactorClientHttpConnector reactorClientHttpConnector = new ReactorClientHttpConnector(client);
-        WebClient build = webClientBuilder.clientConnector(reactorClientHttpConnector)
-                .baseUrl(URL_TRANSACTIONS)
-                .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-        .build();
-        
-        Optional<List<?>> transactionsOptional = Optional.ofNullable(build.method(HttpMethod.GET)
-        .uri(uriBuilder -> uriBuilder
-                .path("/customer/transactions")
-                .queryParam("ibanAccount", ibanAccount)
-                .build())
-        .retrieve()
-        .bodyToFlux(Object.class)
-        .collectList()
-        .block());       
-
-        return transactionsOptional.orElse(Collections.emptyList());
-    }
 }
