+++ paymentchainparent/businessdomain/customer/src/main/java/com/paymentchain/customer/controller/Helpers/CustomerRESTControllerHelper.java
@@ -1,22 +1,110 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package com.paymentchain.customer.controller.Helpers;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.paymentchain.customer.entities.Customer;
+import com.paymentchain.customer.entities.CustomerProduct;
 import com.paymentchain.customer.repository.CustomerRepository;
+import io.netty.channel.ChannelOption;
+import io.netty.channel.epoll.EpollChannelOption;
+import io.netty.handler.timeout.ReadTimeoutHandler;
+import io.netty.handler.timeout.WriteTimeoutHandler;
+import java.time.Duration;
+import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.MediaType;
+import org.springframework.http.client.reactive.ReactorClientHttpConnector;
+import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.reactive.function.client.WebClient;
+import reactor.netty.http.client.HttpClient;
 
-/**
- *
- * @author abela
- */
+@Service
 public class CustomerRESTControllerHelper {
+    private static final String URL_PRODUCTS = "http://BUSINESSDOMAIN-PRODUCT/product";
+    private static final String URL_TRANSACTIONS = "http://localhost:8083/transaction";
     
-    public static Optional<Customer> getCustomerFromRepository(
+    @Autowired
+    private static CustomerRepository customerRepository;
+    
+    @Autowired
+    private static WebClient.Builder webClientBuilder;
+    
+    private static HttpClient client = HttpClient.create()
+            .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, 5000)
+            .option(ChannelOption.SO_KEEPALIVE, true)
+            .option(EpollChannelOption.TCP_KEEPIDLE, 300)
+            .option(EpollChannelOption.TCP_KEEPINTVL, 60)
+            .responseTimeout(Duration.ofSeconds(1))
+            .doOnConnected(connection -> {
+                connection.addHandlerLast(new ReadTimeoutHandler(5000, TimeUnit.MILLISECONDS));
+                connection.addHandlerLast(new WriteTimeoutHandler(5000, TimeUnit.MILLISECONDS));
+            });
+    
+
+            public static Optional<Customer> getCustomerFromRepository(
             CustomerRepository customerRepository,
             long id) {
         return customerRepository.findById(id); 
     }
+    
+    public static Customer getByCode(@RequestParam("code") String code) {
+        Customer customerByCode = customerRepository.findByCode(code);
+        List<CustomerProduct> customerProducts = customerByCode.getProducts();
+        customerProducts.forEach(p -> {
+            String productName = getProductName(p.getId());
+            p.setProductName(productName);
+        });
+        /*
+        String customerIBAN = customerByCode.getIBAN();
+        List<?> transactions = getTransactions(customerIBAN);
+        customerByCode.setTransactions(transactions);
+*/
+       return customerByCode;
+
+    }
+    
+    private static String getProductName(long id) {
+        ReactorClientHttpConnector reactorClientHttpConnector = new ReactorClientHttpConnector(client);
+        WebClient build = webClientBuilder.clientConnector(reactorClientHttpConnector)
+                .baseUrl(URL_PRODUCTS)
+                .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+                .defaultUriVariables(Collections.singletonMap("url", URL_PRODUCTS))
+        .build();
+        
+        JsonNode productResponse = build.method(HttpMethod.GET).uri("/" + id)
+                .retrieve().bodyToMono(JsonNode.class).block();
+        if (productResponse != null) {
+            return productResponse.get("name").asText();
+        }
+        
+        return "";
+    }
+    
+    
+    private List<?> getTransactions(String ibanAccount) {
+        ReactorClientHttpConnector reactorClientHttpConnector = new ReactorClientHttpConnector(client);
+        WebClient build = webClientBuilder.clientConnector(reactorClientHttpConnector)
+                .baseUrl(URL_TRANSACTIONS)
+                .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+        .build();
+        
+        Optional<List<?>> transactionsOptional = Optional.ofNullable(build.method(HttpMethod.GET)
+        .uri(uriBuilder -> uriBuilder
+                .path("/customer/transactions")
+                .queryParam("ibanAccount", ibanAccount)
+                .build())
+        .retrieve()
+        .bodyToFlux(Object.class)
+        .collectList()
+        .block());       
+
+        return transactionsOptional.orElse(Collections.emptyList());
+    }
+    
+    
 }
