+++ paymentchainparent/businessdomain/customer/src/main/java/com/paymentchain/customer/exceptions/BusinessRuleException.java
@@ -1,0 +1,13 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package com.paymentchain.customer.exceptions;
+
+/**
+ *
+ * @author abela
+ */
+public class BusinessRuleException {
+    
+}
